% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRAMkb.R
\name{checkRAMkb}
\alias{checkRAMkb}
\title{Queries available RAM.}
\usage{
checkRAMkb()
}
\value{
Returns a single value.
}
\description{
Queries the operating system and returns the number of kilobytes of memory available.
}
\examples{
checkRAMkb()
}
\author{
Bjorn J. Brooks
}
