% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop2shape.R
\name{crop2shape}
\alias{crop2shape}
\title{Crops and masks a raster to a shapefile.}
\usage{
crop2shape(r, s)
}
\arguments{
\item{r}{A raster to be cropped.}

\item{s}{A shapefile marking the perimeter to crop the raster to.}
}
\value{
Returns a cropped and masked raster.
}
\description{
Crop a raster to the perimeter of a shapefile returning only pixels within shapefile extent. Both raster and shapefile must have same geospatial projection.
}
\author{
Bjorn J. Brooks
}
