% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaTeXtabular.R
\name{LaTeXtabular}
\alias{LaTeXtabular}
\title{Translates R data to LaTeX tabular format}
\usage{
LaTeXtabular(df)
}
\arguments{
\item{df}{R data frame or tabular data.}
}
\value{
Returns the input data with LaTeX code for column breaks and line feeds inserted.
}
\description{
Convert R data or data frame to LaTeX tabular format, suitable for inclusion in an R document.
}
\examples{
# Convert R data output to LaTeX tablular format
cat(LaTeXtabular(mtcars[1:5, 1:5]))
}
\author{
Bjorn J. Brooks
}
